+++ tests/lib/test_storage_capability.py
@@ -42,7 +42,7 @@ class _Stub(Storage):
         raise NotImplementedError
 
     @override
-    def delete(self, key: str) -> None:
+    def delete(self, info: FileInfo) -> None:
         raise NotImplementedError
 
     @override